Fix PDF417 ECI handling, detector, and decoder precision
- Add ECI charset support via eciResult type that tracks current
  encoding and flushes bytes with proper charset conversion
- Fix ECI handling in text compaction, byte compaction, and main
  decode loop to call appendECI() instead of skipping
- Remove non-Java tryHarder logic from detector that caused false
  positives (recursive pattern retry, extra row scanning)
- Fix nil interface trap in detection_result.go and scanning_decoder.go
  where typed nil pointers stored in interfaces appeared non-nil
- Use float32 in codeword decoder sampleBitCounts to match Java float
- Fix stop pattern width scaling in reader min/max codeword width

+++ pdf417/decoder/codeword_decoder.go
@@ -31,14 +31,14 @@ func GetDecodedValue(moduleBitCount []int) int {
 }
 
 func sampleBitCounts(moduleBitCount []int) []int {
-	bitCountSum := sumInts(moduleBitCount)
+	bitCountSum := float32(sumInts(moduleBitCount))
 	result := make([]int, barsInModule)
 	bitCountIndex := 0
 	sumPreviousBits := 0
 	for i := 0; i < modulesInCodeword; i++ {
-		sampleIndex := float64(bitCountSum)/(2.0*float64(modulesInCodeword)) +
-			float64(i)*float64(bitCountSum)/float64(modulesInCodeword)
-		if float64(sumPreviousBits+moduleBitCount[bitCountIndex]) <= sampleIndex {
+		sampleIndex := bitCountSum/(2.0*float32(modulesInCodeword)) +
+			float32(i)*bitCountSum/float32(modulesInCodeword)
+		if float32(sumPreviousBits+moduleBitCount[bitCountIndex]) <= sampleIndex {
 			sumPreviousBits += moduleBitCount[bitCountIndex]
 			bitCountIndex++
 		}
